completed add doctor filter
