+++ front/src/components/HomeComponent/ChatDetails.js
@@ -27,9 +27,6 @@ export default function ChatDetails(props){
         // eslint-disable-next-line
     },[props.getChatData.idDoctor])
 
-    console.log({getChatData:props.getChatData})
-    console.log({getDoctorPrices})
-
     return <div>
         <Typography gutterBottom variant="h6" component="div">
             Solicitar servicios
