+++ front/src/components/HomeComponent/citasPendientes.js
@@ -57,19 +57,24 @@ export default function CitasPendientes(props){
         getUserFunction()
     },[sendRequest])
 
-    console.log({getList})
+    const filterDoctor = getList?.getAllDoctor?.filter(data => {
+        return props.getSecondList?.getAllServices?.some(info => info.idDoctor !== data.id)
+    })
+
+    console.log({filterDoctor})
 
     return auth.rol === "638f3ddd1af87455b52cf7d7" ? <div>
             <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
                 Chatea con mas doctores
             </Typography>
             <div>
-                {getList?.getAllDoctor?.map((index, key) => <ActionAreaCard
+                {filterDoctor && filterDoctor.map((index, key) => <ActionAreaCard
                     key={key}
                     img={index.img}
                     name={index.name}
                     specialty={index.specialty}
                     isLoggedIn={auth.isLoggedIn}
+                    isPacienteLogged={auth.rol === "638f3ddd1af87455b52cf7d7"}
                     onClick={()=>props.onClick(index)}/>)}
             </div>
         </div> :
